+++ packages/core/src/awsService/cloudformation/lsp-server/lspInstaller.ts
@@ -13,6 +13,7 @@ import { getLogger } from '../../../shared/logger/logger'
 import { ResourcePaths } from '../../../shared/lsp/types'
 import * as nodeFs from 'fs' // eslint-disable-line no-restricted-imports
 import globals from '../../../shared/extensionGlobals'
+import { toString } from '../utils'
 
 function determineEnvironment(): CfnLspServerEnvType {
     if (isDebugInstance()) {
@@ -41,7 +42,7 @@ export class CfnLspInstaller extends BaseLspInstaller {
             'awsCfnLsp',
             {
                 resolve: async () => {
-                    const log = getLogger()
+                    const log = getLogger('awsCfnLsp')
                     const cfnManifestStorageKey = 'aws.cloudformation.lsp.manifest'
 
                     try {
@@ -72,7 +73,8 @@ export class CfnLspInstaller extends BaseLspInstaller {
                         throw error
                     }
                 },
-            } as any
+            } as any,
+            'sha256'
         )
     }
 
@@ -95,7 +97,7 @@ export class CfnLspInstaller extends BaseLspInstaller {
         const folders = entries.filter((entry) => entry.isDirectory())
 
         if (folders.length !== 1) {
-            throw new Error(`1 or more CloudFormation LSP folders found ${folders}`)
+            throw new Error(`${folders.length} CloudFormation LSP folders found ${toString(folders)}`)
         }
 
         return {